UIMA-943
fix file deletion - check for null

https://issues.apache.org/jira/browse/UIMA-943

git-svn-id: https://svn.apache.org/repos/asf/incubator/uima/sandbox/trunk@642167 13f79535-47bb-0310-9956-ffa450edef68

+++ DictionaryAnnotator/src/main/java/org/apache/uima/annotator/dict_annot/dictionary/impl/DictionaryCreator.java
@@ -359,8 +359,10 @@ public class DictionaryCreator {
          FileUtils.deleteRecursive(tempDir);
          if (tempDir != null) {
             List files = FileUtils.getFiles(tempDir, true);
-            for (int i = 0; i < files.size(); i++) {
-               ((File) files.get(i)).deleteOnExit();
+            if (files != null) {
+               for (int i = 0; i < files.size(); i++) {
+                  ((File) files.get(i)).deleteOnExit();
+               }
             }
          }
       }
